aggiunto che un user con ruolo_SPECIAL_PLAYER non puoi eliminare/aggiornare tavoli non suoi = bug fix

+++ src/main/java/it/prova/pokeronline/service/tavolo/TavoloServiceImpl.java
@@ -8,10 +8,12 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import it.prova.pokeronline.model.Ruolo;
 import it.prova.pokeronline.model.Tavolo;
 import it.prova.pokeronline.model.Utente;
 import it.prova.pokeronline.repository.tavolo.TavoloRepository;
 import it.prova.pokeronline.repository.utente.UtenteRepository;
+import it.prova.pokeronline.web.api.exception.TavoloNotYourException;
 
 @Service
 public class TavoloServiceImpl implements TavoloService{
@@ -35,12 +37,23 @@ public class TavoloServiceImpl implements TavoloService{
 	@Override
 	@Transactional
 	public void aggiorna(Tavolo tavoloInstance) {
+		String username = SecurityContextHolder.getContext().getAuthentication().getName();
+		Utente utenteInSessionUtente = utenteRepository.findByUsername(username).orElse(null);
+		tavoloInstance.setUtenteCheCreaIlTavolo(utenteInSessionUtente);
+		
 		repository.save(tavoloInstance);
 	}
 
 	@Override
 	@Transactional
 	public void inserisciNuovo(Tavolo tavoloInstance) {
+		String username = SecurityContextHolder.getContext().getAuthentication().getName();
+		Utente utenteInSessionUtente = utenteRepository.findByUsername(username).orElse(null);
+		Tavolo tavoloDaEliminare = caricaSingoloElementoEager(tavoloInstance.getId());
+		if(utenteInSessionUtente.getRuoli().stream().anyMatch(r -> r.getCodice().equals(Ruolo.ROLE_SPECIAL_PLAYER))
+				&& !(utenteInSessionUtente.getId() == tavoloDaEliminare.getUtenteCheCreaIlTavolo().getId()))
+			throw new TavoloNotYourException("impossibile aggiornare un tavolo non creato da te se hai ruolo special player");
+		
 		tavoloInstance.setDataCreazione(LocalDate.now());
 		if(tavoloInstance.getCifraMinima() == null)
 			tavoloInstance.setCifraMinima(0);
@@ -50,6 +63,13 @@ public class TavoloServiceImpl implements TavoloService{
 	@Override
 	@Transactional
 	public void rimuovi(Long idToRemove) {
+		String username = SecurityContextHolder.getContext().getAuthentication().getName();
+		Utente utenteInSessionUtente = utenteRepository.findByUsername(username).orElse(null);
+		Tavolo tavoloDaEliminare = caricaSingoloElementoEager(idToRemove);
+		if(utenteInSessionUtente.getRuoli().stream().anyMatch(r -> r.getCodice().equals(Ruolo.ROLE_SPECIAL_PLAYER))
+				&& !(utenteInSessionUtente.getId() == tavoloDaEliminare.getUtenteCheCreaIlTavolo().getId()))
+			throw new TavoloNotYourException("impossibile eliminare un tavolo non creato da te se hai ruolo special player");
+			
 		repository.deleteById(idToRemove);
 	}
 
@@ -62,8 +82,14 @@ public class TavoloServiceImpl implements TavoloService{
 	public List<Tavolo> listAllByEsperienzaAccumulata() {
 		String username = SecurityContextHolder.getContext().getAuthentication().getName();
 		Utente utenteInSessionUtente = utenteRepository.findByUsername(username).orElse(null);
-		System.err.println(utenteInSessionUtente);
 		return repository.listAllByEsperienzaMinima(utenteInSessionUtente.getEsperienzaAccumulata());
+	}
+
+	@Override
+	public List<Tavolo> listAllByMyCreati() {
+		String username = SecurityContextHolder.getContext().getAuthentication().getName();
+		Utente utenteInSessionUtente = utenteRepository.findByUsername(username).orElse(null);
+		return repository.listAllByMyCreati(utenteInSessionUtente.getId());
 	}
